Fix several circular dependencies
- betwen notation and configuration
- between modeHandler and modeHandlerMap
- between vimState and mode
