+++ src/configuration/notation.ts
@@ -1,5 +1,3 @@
-import { configuration } from './configuration';
-
 export class Notation {
   // Mapping from a regex to the normalized string that it should be converted to.
   private static readonly _notationMap: ReadonlyArray<[RegExp, string]> = [
@@ -79,8 +77,8 @@ export class Notation {
   /**
    * Converts a key to a form which will look nice when logged, etc.
    */
-  public static printableKey(key: string) {
-    const normalized = this.NormalizeKey(key, configuration.leader);
+  public static printableKey(key: string, leaderKey: string) {
+    const normalized = this.NormalizeKey(key, leaderKey);
     return normalized === ' ' ? '<space>' : normalized === '\n' ? '<enter>' : normalized;
   }
